dev/requirements_cloudy_intel.txt
# CloudyIntel Dependencies
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# OpenAI for LLM integration
openai>=1.0.0

# Web search capabilities
google-search-results>=2.4.0

# RAG and vector storage
faiss-cpu>=1.7.0
chromadb>=0.4.0

# Document processing
beautifulsoup4>=4.12.0
requests>=2.32.0
httpx>=0.27.0

# Environment management
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Jupyter and development
jupyter>=1.0.0
ipython>=8.0.0

# Cloud SDKs (optional)
boto3>=1.34.0
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0

# Testing and development
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Additional utilities
typing-extensions>=4.8.0
pydantic>=2.0.0
